make SocketClient static so app only uses one connection
Except for the constructor, all SocketClient functions are now
static, allowing for a single socket to be used for all connections.

This single socket is also managed using a "reference counter" class
called NetActive. On each Activity's onResume the count is incremented;
and when onPause is called the count is decremented. If the count is
zero, then NetDisconnect is called in a sepparate thread to close the
connection. It waits ~2sec and checks if the count is still zero to
finally close the connection.

This allows the app to use a single connection that will auto close
when the app is not visible or not needing a net connection.