update register and login password protocol
The password is send to the server as a SHA-1 hash instead of plain-text.

	SHA1(SHA1(password))

The login protocol has changed to use a very simple zero-knowledge proof.
The server sends a random string on a new connection. The client takes the
double hashed password and appends the random string and hashes it again.
The hash is send to the server, which also does the same. If the hashes
match the client is authenticated.

	SHA1(HashedPassword + RandomString)