relogin after socket disconnects
When the socket disconnected after a successful login, the app still
had the login information but a new login request must be submitted
for each connection. Henceforth, add a new function "relogin" to
NetworkClient that will check the status of SocketClient's new variable
isLoggedin. Each NetworkClient request now also sets loginRequired to
determin if a login is needed. If we're not loggedin and the NetworkClient
request we're running requires a login the submit a new login request.

This also changes the Login page to get the login information after
a successful login from the form fields.