fix crash on losing net connection
If the network socket returns false for isConnected the a new Socket
must be created. Do this by adding a call to hard_disconnect() in
connect() of SocketClient and also removing the check for connection
in hard_disconnect().

NetworkClient and Crypto were also changed so that a lack of connection
can propagate back to the caller of NetworkClient so the user gets
the error message about the lack of internet.