fix clearing of the login EditText when using a hardware keyboad
The setting of the EditText fields in onResume causes them to be
cleared when using a hardware keyboard. This is due to the hardware
keyboard usually causing an oreantation change, forcing a refresh
of the activity.

Instead, set the Login fields in onCreate and update only if the
activity changes by using onActivityResult.