fix deadlock in SyncGameList with no games
the trylock function is used to block multible network requests
from happening simultaneously because NetworkClient.run() creates
a new thread. Unfortunately, if there are no games to sync the class
deadlocks because update_gamestatus never creates a thread.

To fix this, trylock now uses lock++ instead of lock = 1 for setting
the lock test variable. Also a lock-- has been added to update_gamestatus
so the increment/decrements are matched.