fix computer's move when not at top of history
The computer's move can only be applied to the most recent move,
to get there from anywhere in the history call currentMove instead
of forwardMove which only move one move in history.

Rework the use of hindex. Move the updating of hindex into applyMove
and remove it from the history traversing functions forwardMove/currentMove.