fix use of fast gamelist sync when switching user accounts
When logging in as a different user, but with overlaping games, the
game list would be corrupt because only games that had been modified
since the last sync would have the "yourturn" value updated.

Fix this by setting the stime to 0 on all active games after a login.
This would cause all games to be updated on the first sync.