fix strange Java bug when using static refference to class variable
When checking the variable isLoggedin in SocketClient using the
static refference SocketClient.isLoggedin, the value would not always
be correct after a network disconnect. Instead, using an instance
of the class to access the variable gives the correct value, eventhough
the variable is static.

The result was that when a new connection was created, it wouldn't always
send the login request because isLoggedin was set to true.