add notification service and settings activity
GenesisNotifier is a service that checks for new games, msgs or
games which you need to move. It works with GenesisAlarm which
recieves alarm timeouts and starts GenesisNotifier. Henceforth,
the service is only on for short intervals. The simplified algorithm is:

	if not loggedin or service disabled
		quit
	if started from GenesisAlarm
		check server for updates
		send notifications if any
	schedule next wakup using GenesisAlarm
	quit

The Settings activity currently allows for enable/disabling
the polling notifications and the polling frequency.