use android:windowSoftInputMode="adjustPan" for activities with EditText
When the virtual keyboard is activated on enterying an EditText,
the default is to resize the application window. This shrinking
of the window distroys the layout of the activity, especially
when the edittext is above the bottom bar. Instead, use adjustPan
which scrolls the app window to the edittext without resizing.
The default window management setting for