save last sync timestamp in SyncGameList
This fixes a bug where moves get lost if both you and an opponent do
a move while you're in the game board activity. The server now returns
the timestamp on all sync requests. Save the last sync time for games
and messages to the shared preference. On logout, set the last sync to 0.