use a FrameLayout as the root for each OnlineGameList page
For the empty view of the list view to work, the empty view must
be at the same level in the layout hierarchy. The previous solution
put the view with the ViewPager, but that resulted in the view not
sliding with the page changes. The correct solution is to nest each
page in a FrameView, and add the empty view to the frame view.