add "searching for opponent" and GamePoolDialog
Add a TextView above the swipetabs in OnlineGameList which shows
the status of the current user's games he's waiting for (the pool).
Clicking it will bring up a dialog with more information about what
game type and time added. The textview will be hidden if the user
is not waiting for a game.