add UpgradeHandler for app upgrades
The recent change to game syncing using a global timestamp causes
an upgrade issue. Because previous versions don't have the timestamp,
when the user upgrades and opens the online gamelist all archived
games will be added to the your move list.

This is because the last sync time will be zero and a "full" sync
that is done at login is not done which would prevent this from
happening. So add a check for versions <= 27 and log the user out.