GenesisNotifier saves synced data to the database
Before GenesisNotifier only checked the server but didn't save the
new information to the server. This causes a problem with the new
sync time protocol. The yourturn notification would continually go
off for new games that weren't your turn. Now preform the same sync
that is found in SyncGameList and save to the database.