add new high res main menu images
The new main menu images have no words so a TextView must be added to
the layout. Each button consists of a FrameLayout with an ImageView
and a TextView.

The text is resized so that the largest string will fill a column.
This is done by using the new RobotoText.maxTextWidth() function.
It takes a list of strings, a painter and the target width and
returns the optimal pixel font size.

MainMenu.updateLoggedInView() was also added to set the visibility
and text of the welcome message and also hide/show the login button.