fix ACTIVE_SYNC and ARCHIVE_SYNC not performing full resync
A full resync was not done for the active/archive sync types because
the list of gameids we had was removed from the list. Instead, force
downloading all game data. This fixes the bug where there was no way
to get new game field data for games we already had.

Now the resync in Settings will work as intended.