add ExpandablePanel class
ExpandablePanel is a LinearLayout that hides/shows views with a click.
The first child view becomes the header and is the only child shown
on initial view. All other child view have their height set to zero.
When the panel is clicked the view expands to show all views. If
clicked again, the view collapses to only show the header/first view.