fix NetworkClient and SyncClient to run in non-ui thread
NetworkClient.login_user makes a network request, but this is done
on the calling thread (UI) because the function is setting up the
json data before a call to run() on a new thread. Fix by moving the
required network request to login_setup which is called on run().

Similarly, all calls to net.run() not in SocketClient.run() are
run in the UI thread. To fix this create a new NetworkClient
instance and run it on a new thread.