catch NullPointerException in SocketClient.read()
If the server goes down while the socket is open, then input.readLine()
will not return a JSON document. This then causes JSONTokener.nextValue()
to throw a null pointer exception when trying to read a JSON from nothing.
Fix by catching the exception and returning a JSONObject with an error result.