add android:configChanges="orientation" to the OnlineGameList activity
On honeycomb and higher the rotation of the activity still happends
in onCreate before it's locked. The problem is in OnlineGameList
when it does a sync. The dialog comes up and then a rotation causes
the app to crash because the dialog becomes detached. Adding
android:configChanges="orientation" to the manifest prevents the
activity from being destroyed when rotating.