fix crash in Login when finishing the activity with a dialog showing
In the Login activity after the sync, the progress dialog is to be
dismissed before finishing the activity. A crash can happen because
the dialog dismiss code is actually threaded, so finish() on the
activity can happen before the dialog fully closes.

Fix this by adding a Handler sendMessage to ProgressMsg's onStop().
The Handler can be set using a new constructor.