check if game is already archived when updating online game list
When calling updateOnlineGame(), if the game doesn't exist in the
online game table check the archive game table. If the game is
already in the archive table then don't re-add the game back to
the online game table. This fixes the issue where resigned games
would be resynced back to the active game list.