change top and bottom bar to a fixed size using ManualPanel class
In preparation for tablet mode, change the top and bottom bars from
auto scaling to a fixed 50dp. The new bars use an ImageButton. The
background contains the blue color and the src image the contents.

The bar sizeing is done using ManualPanel. Using the "sizes" attr
the child view widths are manually set. For example the following
sizes attr will set the first two views to 1/4 of the total width
and the last view to 1/2.

	sizes="1,1,2/4"