Initial commit to the tablet layout
The tablet UI is a 3 panel layout with a topbar, content and a botbar.
The starting activity is now StartActivity which checks what layout the
user wants using the "tabletMode" boolean preference and loads either
MainMenuPhone or MainMenuTablet.

The phone UI uses a single fragment per activity and is loaded into
the layout at runtime into R.id.fragment01 framelayout using the
BasePhoneActivity class.

SubmitMove is now a dialog. This allows for the UI to be dissabled
except for the SubmitMove. Makeing SubmitMove a fragment would still
allow the user to modify the UI.