improve performance of list adapters by saving data to view tag
Divide the creation of the item view into a setup and reload phase.
The setup phase inflates the view and saves the subviews to a holder
class and then save that to the view tag. The reload phase sets the
data for the view using the holder class. This improves performance
because if the view is already created only the data is updated.