add PawnPromoteDialog to select pawn promotion type
The PawnPromoteDialog is shown in RegGameState.handleMove if the move
was determined to be a promotion. After the dialog is shown the last
value on the callstack poped off. The call to applyMove is done using
the Handler passed to the dialog. A Handler must be used because
Android dialogs are non-modal (nothing can block the UI thread).