add height_policy to ManualPanel to control the child/parent sizes
ManualPanel would set the height for each child view to the size of
the largest child and the ManualPanel to the size requested by the
parent view. This doesn't allow for setting the child views to match
the size of the ManualPanel. Add three height sizing policies:

exact_child:
	set each child to the size it wants to be given the width restrictions
	set the ManualPanel to the size of the largest child view
largest_child:
	set all child views and the ManualPanel to the size of the largest child view
match_parent:
	set all child views and the ManualPanel to the size requested by the parent view

Set the center ManualPanel of the tablet layout to match_parent to fix the
sizing issue where the "No Games" view was not being centered.