remove the move limit on offer draw
The "offer draw" option would only allow sending a draw if the game
had at least 58 moves. This was to force users to only use the draw
option for cases where neither side could get checkmate. But this
restricts several valid use cases:

1) The user accidentally created the wrong game type.
2) The user no longer wants to play the game.
2) Three move repetition, neither side wants to lose pieces.

In each case the only option would be to resign, but that gets
recorded in the user stats -- which a user would not want to do.