add CapturedLayout to show captured pieces below the board
CapturedLayout shows the captured pieces using a row of 11 PieceImg.
The left 5 are captured black pieces, the right 5 are captured white.
There is one empty spacing piece. Highest value pieces are added from
the edge.

The displaying of the captured pieces can be disabled in settings.
PrefKey.SHOW_CAPTURED is the preference key for the setting.