improve Move.setNull() and Move.isNull()
Only one variable in Move needs to be set as Piece.NULL_MOVE and that
is Move.index since a move is invalid if it doesn't refference a valid
index in Board.pieceType. Henceforth, there's no reason to override
setNull; and isNull becomes a check of index for a nagative value.