refactor ObjectArray to take a NewInstance so get() never returns null
changes made in 8c247f618d4221d6a0b017c8423990ab418607a1 cause null
pointer exceptions when running the engine. This is because ObjectArray.get()
was incorrectly assumed to always return a non-null value -- very bad!!!

NewInstance is a new interface with one function getInstance() that
will return a new instance of the class. ObjectArray.get() now checks
for null and creates a new instance. ObjectArray.getNull() can be used
if a null value is required.