load scaled image refferenced in MyImageView on demand
ImageViews hold onto the originally decoded Bitmap from resources to
be able to resize the image when needed. But since the main use case
for MyImageView is to load a static sized scaled down version of an
oversized original, much memory is wasted.

To reduce memory usage and improve performance several changes have
been made. Only the resource id of the image is stored. It can be set
in the constructor using AttributeSet or by calling setResId(int). In
onMeasure() only the image demensions are decoded. Finally, in onDraw()
the scaled image is decoded in a background thread.